+++ src/lib_scoru_wasm/test/test_protocol_migration.ml
@@ -60,7 +60,7 @@ let test_protocol_migration_message ~from_version ~to_version
   Lwt_result_syntax.return_unit
 
 let proto_name : Tezos_scoru_wasm.Pvm_input_kind.protocol -> string = function
-  | Oxford -> "Oxford"
+  | Atlas -> "Atlas"
   | Proto_alpha -> "Proto_alpha"
 
 let tests =
@@ -75,7 +75,7 @@ let tests =
            ~from_version
            ~to_version
            ~after_protocol_activation:protocol))
-    [(V1, V2, Oxford); (V2, V3, Proto_alpha)]
+    [(V1, V2, Atlas); (V2, V3, Proto_alpha)]
 
 let () =
   Alcotest_lwt.run