+++ tezt/long_tests/pipelining.ml
@@ -186,7 +186,7 @@ let init_node_client_with_protocol number_of_additional_bootstrap protocol =
     List.map
       (fun x -> (x, Some 500_000_000, false))
       additional_bootstrap_account
-    (* Starting with Oxford, bootstrap delegates have part of their balance
+    (* Starting with Atlas, bootstrap delegates have part of their balance
        frozen. To avoid having account with none available balance we don't use
        revealed bootstrap account and we reveal them later on. *)
   in